Avanzar relaciones entre entidades

+++ demo/src/main/java/com/example/demo/errors/ProductoNotFoundException.java
@@ -2,7 +2,7 @@ package com.example.demo.errors;
 
 public class ProductoNotFoundException extends RuntimeException {
 
-    public ProductoNotFoundException(int idProducto) {
+    public ProductoNotFoundException(Long idProducto) {
         super("El producto con ID " + idProducto + " no fue encontrado en El Crustáceo Caribeño.");
     }
 }
